+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### v0.17.6
+* Removed a deprecated function which allowed calling non-table attributes on a table.
+* Removed a deprecated function which created an empty table.
+* Fixes bug with copying markers.
+
 ### v0.17.5
 * Eliminated deprecation warnings involved arrays containing arrays/sequences.
 * Changed the column type of a table after a remove operation back to a NumPy array, as written in the documentation.